refactor: uri 관련 메서드 UriHelper 클래스로 이동. RequestParser의 확장자 추출 부분 -> extractExtension HttpRequest의 쿼리 스트링 파싱 부분 -> parseQueryString
